update and delete user

+++ src/pages/api/usuarios/[id].ts
@@ -25,8 +25,8 @@ export default async (req: NextApiRequest, res: NextApiResponse) => {
       }
     case "PUT":
       try {
-        const { name, email, password, rol, id_user } = body;
-        console.log(JSON.stringify(body) + "body 29")
+        const { name, email, password, rol, id_user, new_rol } = body;
+
         let query = 'UPDATE usuarios SET name = $1, email = $2';
         if (!isNilorEmpty(password)) query += ", password = $4 ";
         query += "WHERE id_user = $3 RETURNING *";
@@ -34,21 +34,32 @@ export default async (req: NextApiRequest, res: NextApiResponse) => {
         if (!isNilorEmpty(password)) values.push(password);
         let response = await conn.query(query, values);
 
-        console.log(response + "response 36")
-        query = 'UPDATE usuario_rol SET valido_hasta = $1 WHERE id_user = $2 AND id_rol = $3';
-        values = [rol.valido_hasta, response.rows[0].id_user, rol.id_rol, ];
+        query = 'delete from usuario_rol where id_user = $1 AND id_rol = $2';
+        values = [id_user, rol[0].id_rol]
+        await conn.query(query, values);
+
+        query = 'INSERT INTO usuario_rol (id_user, id_rol, valido_desde, valido_hasta) VALUES ($1, $2, $3, $4)';
+        values = [id_user, new_rol.id, rol[0].valido_desde, new_rol.valido_hasta];
+        console.log(values + "values ")
         response = await conn.query(query, values);
-        // console.log(JSON.stringify(response) + "response 41")
+
         return res.json(response.rows[0]);
       } catch (error: any) {
         return res.status(400).json({ message: error.message });
       }
     case "DELETE":
       try {
-        const text = "DELETE FROM usuarios WHERE id_user = $1 RETURNING *";
-        const values = [id];
-        const result = await conn.query(text, values);
+        let text = "DELETE FROM usuario_rol WHERE id_user = $1 AND id_rol = $2 RETURNING *";
+        let values = [id, body[0].id_rol];
+        let result = await conn.query(text, values);
 
+        if (result.rowCount === 0)
+          return res.status(404).json({ message: "usuario_rol no encontrado" });
+
+        text = "DELETE FROM usuarios WHERE id_user = $1 RETURNING *";
+        values = [id];
+        result = await conn.query(text, values);
+  
         if (result.rowCount === 0)
           return res.status(404).json({ message: "Usuario no encontrado" });
